fix(command_exists): export command_exists

+++ beautify/install.sh
@@ -26,11 +26,11 @@ if [[ ! -d ${plugins}/solarized ]]; then
     if [[ $OS == "osx" ]]; then
         open "${plugins}/solarized/iterm2-colors-solarized/Solarized Dark.itermcolors"
         open "${plugins}/solarized/osx-terminal.app-colors-solarized/xterm-256color/Solarized Dark xterm-256color.terminal"
-    elif [[ $OS == "centos" ]]; then
+    elif [[ $OS == "centos" ]] && command_exists gnome-shell; then
         if [[ ! -d ${plugins}/gnome-terminal-colors-solarized ]]; then
             git clone https://github.com/Anthony25/gnome-terminal-colors-solarized.git
         fi
-        if command_exists; then
+        if command_exists dconf; then
             sudo yum install dconf
         fi
         # need dconf
@@ -43,6 +43,5 @@ if [[ ! -d ${plugins}/dircolors-solarized ]]; then
 	git clone https://github.com/seebi/dircolors-solarized.git
     cp ${plugins}/dircolors-solarized/dircolors.ansi-dark $HOME/.dircolors
 fi
-
 
 exit 0
